chore: refactor chrono custom component and make it responsive
